mobian-tweaks: overide pam.d/common-password to allow numeric passwords
/etc/pam.d/common-password demands a non-numeric and complex password
which is suboptimal as it is used for the lock screen (which for now
can only take digits). So override /etc/pam.d/common-password to alter
the requirements. 4-digit numbers are ok now.

ref: https://wiki.ubuntu.com/PAMConfigFrameworkSpec
Fixes https://gitlab.com/mobian1/issues/-/issues/9

Credits to Mohammed_Sadiq @ purism who implemented this for the Librem5.