tweaks: improve SD card automounting
The initial implementation had a number of issues:
- system partitions (those on the boot devices) were automounted too
- mountpoints names were not really meaningful

In order to solve these issues, this patch introduces a new automount 
script which makes sure we don't mount system partitions, and mounts the 
SD card (or its paritions) on /media/sdcard.

This will make it easier to export the SD card through MTP, which is why 
this patch also includes a new configuration file for umtprd.