tweaks/resize: add BTRFS resize option
In advance of the installer being able to select filesystem, this commit
allows the filesystem being resized to be BTRFS.
BTRFS only allows working with the filesystem when mounted and uses the
mountpoint to reference the filesystem. This is why the code is split into
BTRFS and non-BTRFS sections.

d/control: The initscripts previously relied on e2fsprogs and
f2fs-tools, though this was not listed.