common: split out initramfs and USB gadget scripts
Having a bunch of initramfs scripts and hooks in this package, as well
as USB gadget networking (which doesn't work on all systems) isn't
appropriate for this package and will slow down integration into Debian.

Therefore, they should be dropped and moved to dedicated packages,
namely `mobile-initramfs-tools` and `mobile-usb-networking`.

Similarly, we drop the SSH keygen service (only used when building
custom images, and as such should be converted to an image overlay
in the recipes repo) and ZRAM configuration, as we consider moving to
`systemd-zram-generator` instead of `zram-tools`, which will offer us
better control over how we configure this swap space. In the meantime,
device-specific packages needing it should depend on `zram-tools` and
drop this dependency once we switch to `systemd-zram-generator`.