common: drop polkit policy for ModemManager and add config fragment
We no longer need to use a custom policy file for ModemManager as the
default policy in Debian is now `permissive`, allowing logged in users
to control the modem without additional authentication.

As we're moving to Debian's ModemManager package, we need to enable
the quick suspend/resume feature though, which is easily done by
installing a config fragment for the corresponding systemd service.